Trapped{Data=>Processing}Chain, data-bind / one-run alternatives
When a TrappedProcessingChain contains at least one element
that isExpectingModelData, blackboard connector is added at the beginning
and it is presumed that on the model ->view direction,
the chain is observing data from the model and it is left
to the blackboard connector to start it when data changes.

When a TrappedProcessingChain does not contain any element
that isExpectingModelData, dataTerminator is added at the end
and it is presumed that on the model ->view direction,
the chain is one-time sequence of actions and
it is run immediately with `true` as carried piece of data.